renamed ingest plugin to ingest-geoip plugin, since it only contains the geoip processor
