+++ plugins/ingest-geoip/licenses/geoip2-LICENSE.txt
