+++ plugins/ingest-geoip/licenses/jackson-annotations-LICENSE
