+++ plugins/ingest-geoip/src/main/plugin-metadata/plugin-security.policy
