+++ plugins/ingest-geoip/src/test/java/org/elasticsearch/ingest/geoip/IngestGeoIpRestIT.java
@@ -17,11 +17,10 @@
  * under the License.
  */
 
-package org.elasticsearch.ingest;
+package org.elasticsearch.ingest.geoip;
 
 import com.carrotsearch.randomizedtesting.annotations.Name;
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
-import org.elasticsearch.plugin.ingest.IngestPlugin;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.test.rest.ESRestTestCase;
 import org.elasticsearch.test.rest.RestTestCandidate;
@@ -30,14 +29,14 @@ import org.elasticsearch.test.rest.parser.RestTestParseException;
 import java.io.IOException;
 import java.util.Collection;
 
-public class IngestRestIT extends ESRestTestCase {
+public class IngestGeoIpRestIT extends ESRestTestCase {
 
     @Override
     protected Collection<Class<? extends Plugin>> nodePlugins() {
-        return pluginList(IngestPlugin.class);
+        return pluginList(IngestGeoIpPlugin.class);
     }
 
-    public IngestRestIT(@Name("yaml") RestTestCandidate testCandidate) {
+    public IngestGeoIpRestIT(@Name("yaml") RestTestCandidate testCandidate) {
         super(testCandidate);
     }
