+++ qa/ingest-disabled/build.gradle
@@ -19,13 +19,8 @@
 
 apply plugin: 'elasticsearch.rest-test'
 
-dependencies {
-    testCompile project(path: ':plugins:ingest', configuration: 'runtime')
-}
-
 integTest {
     cluster {
-        plugin 'ingest', project(':plugins:ingest')
         systemProperty 'es.node.ingest', 'false'
     }
 }
