+++ qa/ingest-with-mustache/build.gradle
@@ -20,12 +20,5 @@
 apply plugin: 'elasticsearch.rest-test'
 
 dependencies {
-    testCompile project(path: ':plugins:ingest', configuration: 'runtime')
     testCompile project(path: ':modules:lang-mustache', configuration: 'runtime')
-}
-
-integTest {
-    cluster {
-        plugin 'ingest', project(':plugins:ingest')
-    }
 }
